file_cleaning
Obtained data from Stack Overflow API, with tag 'ggplot2' and 'R'. File contains three features: score of the question (number of upvotes - number of downvotes), title of the question, and body description (in HTML format). The score was turned into a factor variable. For instance, if the score was greater than 0, the question was interpreted useful. If not, the question was interpreted as not useful.  

The other two variables were cleaned separately before combining them into the same file for modeling purposes. The cleaning process includes extracting the HTML format using the tm.plugin.webmining library.