model_build
Modeled the data using Random Forest (Ranger), CART, Logistic Regression, and Linear Discriminant Analysis. Created ROC Curves for the Logistic Regression and LDA models. Used bootstrap to assess the 95% Confidence Interval of our model performance estimates: Accuracy, TPR, and FPR. 

Found that if a Stack Overflow manager wanted to maximize the number of useful questions on the front page rather than have a baseline approach that randomizes the questions (Useful or not), she could use the model technique that lowers the False Positive Rate the most optimally. That is, she should use the logistic regression model, the model with the lowest False Positive Rate.